+++ Source/ModuleFileSystem.h
@@ -21,7 +21,7 @@ public:
 	bool Exists(const char* file) const;
 	bool CreateDir(const char* dir);
 	bool IsDirectory(const char* file) const;
-	int LastModificationTime(const char* file)const;
+	signed long long LastModificationTime(const char* file)const;
 	void DiscoverFiles(const char* directory, std::vector<std::string>& file_list, std::vector<std::string>& dir_list) const;
 	bool FileExists(const char* file);
 	void SplitFilePath(const char* full_path, std::string* path, std::string* file = nullptr, std::string* extension = nullptr) const;
